+++ emubase/Defines.h
@@ -15,7 +15,7 @@ UKNCBTL. If not, see <http://www.gnu.org/licenses/>. */
 
 //////////////////////////////////////////////////////////////////////
 
-inline int GetDigit (uint16_t word, int pos)
+inline uint16_t GetDigit(uint16_t word, int pos)
 {
     //return (word >>= pos * 3) % 8;
     return (word >>= ((pos << 1) + pos)) & 7;
